[SWM-206] Refactor : util directory
