+++ src/api/lectures/search.ts
@@ -1,7 +1,7 @@
-import getQueryURL from '@/src/util/getQueryURL';
+import getQueryURL from '@/src/util/http/getQueryURL';
 import { Endpoints } from '../Endpoints';
 import { ErrorMessage } from '../ErrorMessage';
-import { getAuthorizedHeaders } from '@/src/util/getAuthorizedHeaders';
+import { getAuthorizedHeaders } from '@/src/util/http/getAuthorizedHeaders';
 
 export async function fetchLecturesByKeyword(params: SearchLectureParams) {
   const headers = await getAuthorizedHeaders();
