+++ src/util/error/getErrorObject.ts
