--- src/util/getRandomID.ts
@@ -1,3 +1,0 @@
-export default function getRandomID() {
-  return String(new Date().getTime());
-}
